+++ Public Key/RSA_KeyGenerator.java
@@ -1,16 +1,16 @@
 
 public class RSA_KeyGenerator extends KeyGenerator
 {
-    public int[][] generateKeys(int[] pq)
+    public long[][] generateKeys(long[] pq)
     {
-        int p = pq[0];
-        int q = pq[1];
-        int n = p*q;
-        int t = (p-1)*(q-1);
-        int e = 17;
-        float dTemp = 1; // arbitrary such that its not an int also math class is weird and won't round doubles to ints
-        float z = 0;
-        int d = 0;
+        long p = pq[0];
+        long q = pq[1];
+        long n = p*q;
+        long t = (p-1)*(q-1);
+        long e = 17;
+        double dTemp = 1; // arbitrary such that its not an int also math class is weird and won't round doubles to ints
+        double z = 0;
+        long d = 0;
         while( Math.abs((dTemp - d)) > .00001)//check if dTemp is an int
         {
             z++;
@@ -18,9 +18,9 @@ public class RSA_KeyGenerator extends KeyGenerator
             d = Math.round(dTemp);
         }
         
-        int[] publicKey = {n,e};
-        int[] privateKey = {n,d};
-        int[][] keys = {publicKey,privateKey};
+        long[] publicKey = {n,e};
+        long[] privateKey = {n,d};
+        long[][] keys = {publicKey,privateKey};
         return keys;
     }
 }
